first impressions of directional light
the way I added padding creates seams between surfaces - the original simple method of simply making the light maps 1 unit bigger everywhere is better because it forces the calculation of light in the virtual patch as if the surface was extending there. this means linear filtering looks more natural with no (or visibly less? idk have to try it) seams.